+++ lib/fisk8_result_viewer/score/updater.rb
@@ -2,15 +2,17 @@ module Fisk8ResultViewer
   module Score
     class Updater
       include Skater::FindSkater
-      def update_scores(url, competition, category, segment, parser:)
+      def update_scores(url, competition, category, segment, parser:, attributes: {})
         parser.parse_scores(url).each do |parsed_score|
-          update_score(parsed_score, competition, category, segment)
+          update_score(parsed_score, competition, category, segment, attributes: attributes)
         end
       end
 
-      def update_score(parsed_score, competition, category, segment)
-        keys = [:skater_name, :ranking, :tss, :tes, :pcs, :deductions]
+      def update_score(parsed_score, competition, category, segment, attributes: {})
+        keys = [:skater_name, :ranking, :nation, :starting_number, :tss, :tes, :pcs, :deductions, :deduction_reasons, :result_pdf, :base_value]
         competition.scores.create!(parsed_score.slice(*keys)) do |score|
+          score.competition_name = competition.name
+          score.attributes = attributes
           ## category_ results
           parsed_score[:skater_name] = correct_skater_name(parsed_score[:skater_name])
           cr = find_relevant_category_result(competition.category_results.where(category: category), parsed_score[:skater_name], segment, parsed_score[:ranking]) || raise("no such skater: '#{skater_name}' in #{category}")
@@ -34,12 +36,12 @@ module Fisk8ResultViewer
           score[:elements_summary] = parsed_score[:elements].map do |element|
             keys = [:number, :name, :info, :base_value, :credit, :goe, :judges, :value]
             score.elements.create!(element.slice(*keys)).name
-          end
+          end.join(',')
           ## components
           score[:components_summary] = parsed_score[:components].map do |component|
             keys = [:number, :name, :factor, :judges, :value]
             score.components.create!(component.slice(*keys)).value
-          end
+          end.join(',')
 
           puts "    %s-%s [%2d] %-40s (%6d) | %6.2f = %6.2f + %6.2f + %2d" % [score.category, score.segment, score.ranking, score.skater_name, score.skater.isu_number.to_i, score.tss.to_f, score.tes.to_f, score.pcs.to_f, score.deductions.to_i]
         end